:recycle: createstorage: Inline function
