+++ src/cutty/filestorage/adapters/cookiecutter.py
@@ -1,10 +1,30 @@
 """File storage for Cookiecutter projects."""
 import pathlib
 from collections.abc import Sequence
+from typing import Optional
 
+from cutty.filestorage.adapters.disk import DiskFileStorage
+from cutty.filestorage.adapters.disk import FileExistsPolicy
+from cutty.filestorage.adapters.observers.cookiecutter import CookiecutterHooksObserver
+from cutty.filestorage.adapters.observers.git import GitRepositoryObserver
 from cutty.filestorage.domain.files import File
+from cutty.filestorage.domain.observers import FileStorageObserver
+from cutty.filestorage.domain.observers import observe
 from cutty.filestorage.domain.storage import FileStorage
 from cutty.filesystems.domain.path import Path
+
+
+def fileexistspolicy(
+    overwrite_if_exists: bool, skip_if_file_exists: bool
+) -> FileExistsPolicy:
+    """Return the policy for overwriting existing files."""
+    return (
+        FileExistsPolicy.RAISE
+        if not overwrite_if_exists
+        else FileExistsPolicy.SKIP
+        if skip_if_file_exists
+        else FileExistsPolicy.OVERWRITE
+    )
 
 
 def createcookiecutterstorage(
@@ -15,8 +35,19 @@ def createcookiecutterstorage(
     hookfiles: Sequence[File],
 ) -> FileStorage:
     """Create storage for Cookiecutter project files."""
-    from cutty.services.create import createstorage
+    fileexists = fileexistspolicy(overwrite_if_exists, skip_if_file_exists)
+    storage: FileStorage = DiskFileStorage(project_dir.parent, fileexists=fileexists)
 
-    return createstorage(
-        template_dir, project_dir, overwrite_if_exists, skip_if_file_exists, hookfiles
-    )
+    observer: Optional[FileStorageObserver] = None
+
+    if hookfiles:  # pragma: no branch
+        observer = CookiecutterHooksObserver(
+            hookfiles=hookfiles, project=project_dir, fileexists=fileexists
+        )
+
+    if observer:  # pragma: no branch
+        storage = observe(storage, observer)
+
+    storage = observe(storage, GitRepositoryObserver(project=project_dir))
+
+    return storage
